GSYE-297: Move forward profile generation to separate module
